refactor: moved User profile pages and components into single module folder
