+++ src/front/modules/user-profile/user-profile-pages/UserProfile.jsx
@@ -17,9 +17,9 @@ import {
   Lock,
   MessageCircle
 } from "lucide-react";
-import { EditDescriptionModal } from "../components/EditDescriptionModal";
-import { COUNTRIES, CountryFlag } from "../shared/utils/country-flag-name-display/countriesData.jsx";
-import CollectionCard from "../modules/collections/collections-subcomponents/CollectionCard.jsx";
+import { EditDescriptionModal } from "../user-profile-subcomponents/EditDescriptionModal.jsx";
+import { COUNTRIES, CountryFlag } from "../../../shared/utils/country-flag-name-display/countriesData.jsx";
+import CollectionCard from "../../collections/collections-subcomponents/CollectionCard.jsx";
 
 export const UserProfile = () => {
   const { username } = useParams();
